Fix bug in final_spin_from_initial

+++ pycbc/conversions.py
@@ -1215,8 +1215,8 @@ def get_final_from_initial(mass1, mass2, spin1x=0., spin1y=0., spin1z=0.,
     final_mass = numpy.zeros(mass1.shape)
     final_spin = numpy.zeros(mass1.shape)
     for ii in range(final_mass.size):
-        m1 = mass1[ii]
-        m2 = mass2[ii]
+        m1 = numpy.float(mass1[ii])
+        m2 = numpy.float(mass2[ii])
         spin1 = [spin1x[ii], spin1y[ii], spin1z[ii]]
         spin2 = [spin2x[ii], spin2y[ii], spin2z[ii]]
         _, fm, fs = lalsim.SimIMREOBFinalMassSpin(m1, m2, spin1, spin2,
